_ggc script:fix bug deleting previously deleted profile.
Say profile some-profile was created and then deleted. Then we
recreate some-profile and then try to delete - we will get an
error since we do not actually delete the profile but move it into
tmp/. And since the name already exists in /tmp, the `mv`
operation will fail. This patch suffixes the name with the
current timestamp to make the name unique inside /tmp.