Again! Handle quoting in script p-env.
It turns out when using --env-file argument in docker/compose,
quoted strings are quoted again (for some reason, this does not
happen when using compose in elixir project in dev - quotes are
never re-quoted). So if the output format is sh, we do not handle
quoting specially - we rely on caller to use proper quoting
before calling this script. For json/yaml, we handle quoting in
a manner appropriate to those formats.