c-tmux: allow user to change session state in resurrect file.
When tmux is launched with a session that is not the state in resurrect
file, the state in resurrect file will be loaded automatically by the
resurrect plugin. We will then have to manually switch to the session we
gave to c-tmux. This PR ensures that session given as argument to c-tmux
is always loaded and manually switching is not required.